MSI: 343 Implement show stock qty for product
