+++ app/code/Magento/InventoryCatalog/view/adminhtml/ui_component/product_listing.xml
@@ -17,9 +17,9 @@
                 <label translate="true">Quantity</label>
             </settings>
         </column>
-        <column name="qty_stock"
-                class="Magento\InventoryCatalog\Ui\DataProvider\Component\Listing\Column\Stock"
-                component="Magento_InventoryCatalog/js/product/grid/cell/stock"
+        <column name="stocks"
+                class="Magento\InventoryCatalog\Ui\Component\Listing\Column\Stocks"
+                component="Magento_InventoryCatalog/js/product/grid/cell/stocks"
                 sortOrder="85">
             <settings>
                 <filter>false</filter>
