--- app/code/Magento/InventoryCatalog/view/adminhtml/web/template/product/grid/cell/stock-cell.html
@@ -1,15 +1,0 @@
-<!--
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
--->
-<div class="data-grid-cell-content">
-    <ul data-bind="foreach: { data: $col.getStockData($row()), as: '$sourceItemData' }">
-        <li>
-            <strong data-bind="text: $sourceItemData.stock_name"></strong>
-            :
-            <span data-bind="text: $sourceItemData.qty"></span>
-        </li>
-    </ul>
-</div>
