+++ app/code/Magento/InventoryCatalog/view/adminhtml/web/template/product/grid/cell/stocks-cell.html
@@ -1,0 +1,15 @@
+<!--
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+-->
+<div class="data-grid-cell-content">
+    <ul data-bind="foreach: { data: $col.getStockData($row()), as: '$stockData' }">
+        <li>
+            <strong data-bind="text: $stockData.stock_name"></strong>
+            :
+            <span data-bind="text: $stockData.qty"></span>
+        </li>
+    </ul>
+</div>
